added dispaly for window without masterclasses
