+++ frontend/src/pages/MyMasterclassesPage/MyMasterclassesPage.module.scss
@@ -1,0 +1,73 @@
+@use "../../utils/variables/global.scss";
+
+.text {
+  font-family: var(--font-primary);
+  font-size: 20px;
+  font-weight: 400;
+  line-height: 1.35;
+}
+
+.link {
+  color: var(--text-color);
+  text-decoration: underline;
+}
+
+.flexContainer {
+  margin-top: 70px;
+  display: flex;
+  gap: 40px;
+}
+
+.calendarContainer {
+  flex: 1; 
+  max-width: 510px;
+}
+
+.masterclassContainer {
+  flex: 2; 
+}
+
+.title {
+  margin-bottom: 48px;
+  font-family: var(--font-primary);
+  font-size: 32px;
+  font-weight: 700;
+  line-height: 1.35;
+}
+
+.list {
+  margin: 40px 0;
+  width: 730px;
+  display: flex;
+  justify-content: left;
+  align-items: baseline;
+  list-style-type: none;
+  border-bottom: 1px solid var(--color-green-grey);
+}
+
+.wrapper {
+  margin-top: 40px; 
+}
+
+.subtitle {
+  margin: 22px 0 16px 0;
+  font-family: var(--font-primary);
+  font-size: 32px;
+  font-weight: 700;
+  line-height: 1.35;
+}
+
+.cardMoreContentContainer {
+  display: flex; 
+  gap: 40px; 
+}
+
+
+  @media screen and (max-width: global.$breakpoint-mobile) {
+  
+    .list {
+      min-width: 100%;
+      padding: 0;
+      padding-top: 8px;
+    }
+  }
