+++ frontend/src/pages/MyMasterclassesPastPage/MyMasterclassesPastPage.jsx
@@ -1,11 +1,12 @@
 import { classNames } from '../../helpers/classNames';
 import cls from './MyMasterclassesPastPage.module.scss';
 import Calendar from '../../components/Calendar/Calendar';
-import { MyMasterclassesHeaderLinks } from '../../components/MyMasterclassesHeaderLink/MyMasterclassesHeaderLinks';
+import { MyMasterclassesHeaderLinks } from '../../components/MyMasterclassesHeaderLinks/MyMasterclassesHeaderLinks';
 import { masterclasses } from '../../utils/consts/mockMasterclasses';
 import MasterclassList from '../../components/MasterclassList/MasterclassList';
 
 export const MyMasterclassesPastPage = () => {
+  const message = 'Нет прошедших занятий';
   return (
     <div className={cls.flexContainer}>
       <div className={cls.calendarContainer}>
@@ -16,7 +17,7 @@ export const MyMasterclassesPastPage = () => {
         <ul className={classNames(cls.list, {}, [])}>
           <MyMasterclassesHeaderLinks />
         </ul>
-        <MasterclassList masterclasses={masterclasses} />
+        <MasterclassList masterclasses={masterclasses} message={message} />
       </div>
     </div>
   );
