+++ frontend/src/utils/consts/mockMasterclasses.js
@@ -1,58 +1,58 @@
 export const masterclasses = [
-  {
-    // eslint-disable-next-line max-len
-    description: 'Мастер-класс по римской мозаике однодневный, дополнительнйы, в группе в мастером Соколовым  Антоном и Маргаритой Травушкиной',
-    date: '13 июня',
-    time: '17:00',
-    duration: '3,5 часа',
-    price: '7 000',
-    teacher: 'Анна Павлова',
-    pay: true,
-  },
-  {
-    description: 'Мастер-класс по римской мозаике, однодневный',
-    date: '20 июня',
-    time: '18:00',
-    duration: '4 часа',
-    price: '8 000',
-    teacher: 'Иван Иванов',
-    pay: false,
-  },
-  {
-    // eslint-disable-next-line max-len
-    description: 'Мастер-класс по римской мозаике однодневный, дополнительнйы, в группе в мастером Соколовым  Антоном и Маргаритой Травушкиной',
-    date: '27 июня',
-    time: '19:00',
-    duration: '3 часа',
-    price: '6 500',
-    teacher: 'Мария Сидорова',
-    pay: true,
-  },
-  {
-    description: 'Мастер-класс по римской мозаике, однодневный',
-    date: '4 июля',
-    time: '16:00',
-    duration: '2,5 часа',
-    price: '5 500',
-    teacher: 'Петр Петров',
-    pay: false,
-  },
-  {
-    description: 'Мастер-класс по римской мозаике, однодневный',
-    date: '11 июля',
-    time: '15:00',
-    duration: '3 часа',
-    price: '7 500',
-    teacher: 'Анна Павлова',
-    pay: true,
-  },
-  {
-    description: 'Мастер-класс по римской мозаике, однодневный',
-    date: '18 июля',
-    time: '14:00',
-    duration: '4 часа',
-    price: '8 500',
-    teacher: 'Ольга Смирнова',
-    pay: false,
-  },
+  // {
+  //   // eslint-disable-next-line max-len
+  //   description: 'Мастер-класс по римской мозаике однодневный, дополнительнйы, в группе в мастером Соколовым  Антоном и Маргаритой Травушкиной',
+  //   date: '13 июня',
+  //   time: '17:00',
+  //   duration: '3,5 часа',
+  //   price: '7 000',
+  //   teacher: 'Анна Павлова',
+  //   pay: true,
+  // },
+  // {
+  //   description: 'Мастер-класс по римской мозаике, однодневный',
+  //   date: '20 июня',
+  //   time: '18:00',
+  //   duration: '4 часа',
+  //   price: '8 000',
+  //   teacher: 'Иван Иванов',
+  //   pay: false,
+  // },
+  // {
+  //   // eslint-disable-next-line max-len
+  //   description: 'Мастер-класс по римской мозаике однодневный, дополнительнйы, в группе в мастером Соколовым  Антоном и Маргаритой Травушкиной',
+  //   date: '27 июня',
+  //   time: '19:00',
+  //   duration: '3 часа',
+  //   price: '6 500',
+  //   teacher: 'Мария Сидорова',
+  //   pay: true,
+  // },
+  // {
+  //   description: 'Мастер-класс по римской мозаике, однодневный',
+  //   date: '4 июля',
+  //   time: '16:00',
+  //   duration: '2,5 часа',
+  //   price: '5 500',
+  //   teacher: 'Петр Петров',
+  //   pay: false,
+  // },
+  // {
+  //   description: 'Мастер-класс по римской мозаике, однодневный',
+  //   date: '11 июля',
+  //   time: '15:00',
+  //   duration: '3 часа',
+  //   price: '7 500',
+  //   teacher: 'Анна Павлова',
+  //   pay: true,
+  // },
+  // {
+  //   description: 'Мастер-класс по римской мозаике, однодневный',
+  //   date: '18 июля',
+  //   time: '14:00',
+  //   duration: '4 часа',
+  //   price: '8 500',
+  //   teacher: 'Ольга Смирнова',
+  //   pay: false,
+  // },
 ];
